test-requirements.txt
flake8
tox
nose==1.3.7
pyyaml==3.13